+++ src/Repository/Contract/AbstractRepositoryContract.php
@@ -6,7 +6,8 @@ interface AbstractRepositoryContract
 {
     public function find(int $id);
     public function findOneWhere(array $criteria);
-    public function findAll();
+    public function findAll(): array;
+    public function findAllWhere(array $criteria, ?array $orderBy = null, $limit = null, $offset = null): array;
     public function store(array $data);
     public function update(object $entity, array $data);
 }