+++ src/Repository/Contract/TaskRepositoryContract.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Repository\Contract;
+
+interface TaskRepositoryContract extends AbstractRepositoryContract {}